Document array/eval use
The backup.sh script we are calling uses mongodump, which puts info on stderr and is not captured by a variable when executed using $(), and when using sudo/su, redirection is very tricky to use.  So, to prevent the output from appearing when not VERBOSE=TRUE, we use eval to make redirecting stderr simpler.  We use an array for the command line to ensure proper variable and parameter quoting and separation.